refactor: rename features dir to modules
