+++ src/modules/auth/auth.router.ts
@@ -1,6 +1,6 @@
 import Router from '@koa/router';
 import { object, string, ref } from 'yup';
-import { AuthController } from '@/features/auth/auth.controller';
+import { AuthController } from '@/modules/auth/auth.controller';
 import { vaidateRequestBodyMiddleware } from '@/common/middlewares/validation/validate-request-body.middleware';
 import { ISignInDto } from '@/interfaces/dtos/auth/sign-in.dto';
 import { ISignUpDto } from '@/interfaces/dtos/auth/sign-up.dto';
