+++ src/modules/auth/utils/generate-token.ts
