+++ src/modules/users/middlewares/find-user-by-id.middleware.ts
