+++ src/modules/users/users.router.ts
@@ -1,9 +1,9 @@
 import Router from '@koa/router';
 import { object, string } from 'yup';
-import { UsersController } from '@/features/users/users.controller';
+import { UsersController } from '@/modules/users/users.controller';
 import { vaidateRequestBodyMiddleware } from '@/common/middlewares/validation/validate-request-body.middleware';
 import { vaidateRequestParamsMiddleware } from '@/common/middlewares/validation/validate-request-params.middleware';
-import { findUserByIdMiddleware } from '@/features/users/middlewares/find-user-by-id.middleware';
+import { findUserByIdMiddleware } from '@/modules/users/middlewares/find-user-by-id.middleware';
 import { isCurrentUserMiddleware } from '@/common/middlewares/authorization/is-current-user.middleware';
 import { IUpdateUserDto } from '@/interfaces/dtos/users/update-user.dto';
 import { objectId } from '@/common/yup/custom-schemas/object-id.schema';
